% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itunes.R
\docType{package}
\name{itunes}
\alias{itunes}
\alias{itunes-package}
\title{Access iTunes Top Hits, Movies, Apps and their associated reviews}
\description{
Access iTunes Top Hits, Movies, Apps and their associated reviews
}
\author{
Zeyad Deeb (zeyad.deeb@icloud.com.com)
}
