% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-itunes.R
\name{search_itunes}
\alias{search_itunes}
\title{Search for an Podcast, App id on iTunes}
\usage{
search_itunes(searchterm = NULL)
}
\arguments{
\item{searchterm}{}
}
\value{
data.frame
}
\description{
Search for an Podcast, App id on iTunes
}
\examples{
\dontrun{
top_stuff <- search_itunes('Radiolab')
}
}
