% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-top.R
\name{get_top}
\alias{get_top}
\title{Get Top n Apps, Songs, Albums, Movies and more... from iTunes}
\usage{
get_top(n = 10)
}
\arguments{
\item{media.types}{defaults to all current media types}
}
\value{
Data Frame
}
\description{
This function returns the id(s) for the top 10 apps, that can be used to
pull reviews and ratings for a one or more apps
}
